🙈 errors

+++ app/cogs/imageboard.py
@@ -1,6 +1,8 @@
 import discord
 from discord.ext import commands
 from data import db
+import re
+from utils import permissions, http
 
 
 class ImageBoard(commands.Cog):
@@ -12,6 +14,7 @@ class ImageBoard(commands.Cog):
 
     @commands.group(name='img', aliases=['i'])
     async def img(self, ctx):
+        """ Image Board """
         if ctx.invoked_subcommand is None:
             await ctx.message.delete()
             await ctx.send('You\'re missing an argument 😥\nSee `-help img`')
@@ -27,30 +30,49 @@ class ImageBoard(commands.Cog):
 
     @displayImg.error
     async def displayImg_error(self, ctx, error):
-        await ctx.message.delete()
+        # await ctx.message.delete()
         if isinstance(error, commands.MissingRequiredArgument):
             await ctx.send('You\'re missing an argument 😥\nSee `-help img show`')
 
     @img.command(name='add', aliases=['a'])
+    @commands.check(permissions.hasManagerRole)
     async def addImg(self, ctx, *, name):
-        keyname = name.lower().replace(" ", "-")
-        whiteList = ['jpeg', 'jpg', 'webp', 'png', 'gif']
-
-        att = ctx.message.attachments[0]
-        url = att.url
-        if att.filename.split('.')[-1] not in whiteList:
-            await ctx.send(att.filename.split('.')[-1]+" is not a valid image format")
-        else:
-            db.saveTo(ctx.guild.id, 'imageboard', (keyname, url))
+        whiteList = ['.jpeg', '.jpg', '.webp', '.png', '.gif']
+        filename = re.sub(
+            r'http\S+', '', name).strip().lower().replace(" ", "-")
+        url = re.search("(?P<url>https?://[^\s]+)", name)
+        if url:
+            await ctx.message.delete()
+            error = http.fileIsType(url.group("url"), whiteList)
+            if error:
+                return await ctx.send(f"**{error}** is not a valid type")
+            saveerror = db.saveTo(ctx.guild.id, 'imageboard',
+                                  filename, url.group("url"))
+            if saveerror:
+                return await ctx.send(f"**{filename}** already exists 🤔")
+            await ctx.send("☁☁ Done!")
+        elif ctx.message.attachments:
+            url = ctx.message.attachments[0].url
+            error = http.fileIsType(url, whiteList)
+            if error:
+                await ctx.message.delete()
+                return await ctx.send(f"**{error}** is not a valid type")
+            saveerror = db.saveTo(ctx.guild.id, 'imageboard', filename, url)
+            if saveerror:
+                await ctx.message.delete()
+                return await ctx.send(f"**{filename}** already exists 🤔")
             await ctx.send("⬆ Don't delete the source image\n☁☁ Done!")
+        else:
+            await ctx.message.delete()
+            await ctx.send("No URL or file 👎")
 
-    @addImg.error
+    @ addImg.error
     async def addImg_error(self, ctx, error):
-        await ctx.message.delete()
+        # await ctx.message.delete()
         if isinstance(error, commands.MissingRequiredArgument):
             await ctx.send('You\'re missing an argument 😥\nSee `-help img add`')
 
-    @img.command(name='list', aliases=['l'])
+    @ img.command(name='list', aliases=['l'])
     async def listImg(self, ctx):
         await ctx.message.delete()
         imglist = db.listKeysFrom(ctx.guild.id, 'imageboard')
@@ -60,15 +82,16 @@ class ImageBoard(commands.Cog):
             embed.add_field(name=name, value="\u200b")
         message = await ctx.channel.send(embed=embed)
 
-    @img.command(name='remove', aliases=['rm'])
+    @ img.command(name='remove', aliases=['rm'])
+    @ commands.check(permissions.hasManagerRole)
     async def removeImg(self, ctx, name):
         await ctx.message.delete()
         db.delKey(ctx.guild.id, 'imageboard', name)
         await ctx.send('🧯 Done!')
 
-    @removeImg.error
+    @ removeImg.error
     async def removeImg_error(self, ctx, error):
-        await ctx.message.delete()
+        # await ctx.message.delete()
         if isinstance(error, commands.MissingRequiredArgument):
             await ctx.send('You\'re missing an argument 😥\nSee `-help img remove`')
